requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
discogs-client==2.2.1
Django==1.10.3
djangorestframework==3.5.3
EasyProcess==0.2.3
httpie==0.9.9
imgurpython==1.1.7
oauthlib==2.0.1
Pillow==3.4.2
Pygments==2.1.3
PyVirtualDisplay==0.2.1
requests==2.12.1
selenium==3.0.2
six==1.10.0
